moving of utility functions to io module
