+++ pylinac/core/io.py
@@ -6,7 +6,54 @@ from urllib.error import HTTPError, URLError
 from urllib.request import urlretrieve, urlopen
 import zipfile
 
+import dicom
 from tqdm import tqdm
+
+
+def is_dicom(file):
+    """Boolean specifying if file is a proper DICOM file.
+
+    This function is a pared down version of read_preamble meant for a fast return.
+    The file is read for a proper preamble ('DICM'), returning True if so,
+    and False otherwise. This is a conservative approach.
+
+    Parameters
+    ----------
+    file : str
+        The path to the file.
+
+    See Also
+    --------
+    pydicom.filereader.read_preamble
+    pydicom.filereader.read_partial
+    """
+    with open(file, 'rb') as fp:
+        fp.read(0x80)
+        prefix = fp.read(4)
+        return prefix == b"DICM"
+
+
+def is_dicom_image(file):
+    """Boolean specifying if file is a proper DICOM file with a image
+
+    Parameters
+    ----------
+    file : str
+        The path to the file.
+
+    See Also
+    --------
+    pydicom.filereader.read_preamble
+    pydicom.filereader.read_partial
+    """
+    result = False
+    try:
+        img = dicom.read_file(file, force=True)
+        img.pixel_array
+        result = True
+    except:
+        pass
+    return result
 
 
 def is_zipfile(file):
