+++ pylinac/core/utilities.py
@@ -68,15 +68,6 @@ def is_close(val, target, delta=1):
     return False
 
 
-def import_mpld3():
-    """Try importing MPLD3. Raises error if not installed. Returns the MPLD3 library."""
-    try:
-        import mpld3
-    except ImportError:
-        raise ImportError("The MPLD3 library must be installed to make interactive plots. See http://mpld3.github.io/index.html for info.")
-    return mpld3
-
-
 def typed_property(name, expected_type_or_tuple_of_types):
     """Type-enforced property. Python Cookbook 9.21 (3rd ed)."""
     storage_name = '_' + name
@@ -99,52 +90,6 @@ def simple_round(number, decimals=0):
     num = int(round(number * 10 ** decimals))
     num /= 10 ** decimals
     return num
-
-
-def is_dicom(file):
-    """Boolean specifying if file is a proper DICOM file.
-
-    This function is a pared down version of read_preamble meant for a fast return.
-    The file is read for a proper preamble ('DICM'), returning True if so,
-    and False otherwise. This is a conservative approach.
-
-    Parameters
-    ----------
-    file : str
-        The path to the file.
-
-    See Also
-    --------
-    pydicom.filereader.read_preamble
-    pydicom.filereader.read_partial
-    """
-    fp = open(file, 'rb')
-    preamble = fp.read(0x80)
-    prefix = fp.read(4)
-    return prefix == b"DICM"
-
-
-def is_dicom_image(file):
-    """Boolean specifying if file is a proper DICOM file with a image
-
-    Parameters
-    ----------
-    file : str
-        The path to the file.
-
-    See Also
-    --------
-    pydicom.filereader.read_preamble
-    pydicom.filereader.read_partial
-    """
-    result = False
-    try:
-        img = dicom.read_file(file, force=True)
-        img.pixel_array
-        result = True
-    except:
-        pass
-    return result
 
 
 def isnumeric(object):
